+++ __init__.py
@@ -25,8 +25,8 @@ bl_info = {
     "blender": (2, 93, 0),
     "location": "Properties > Render",
     "description": "Saves presets of all settign render tab",
-    "warning": "https://github.com/schroef/Save-Render-Presets/issues",
-    "doc_url": "https://github.com/schroef/Save-Render-Presets",
+    "tracker_url": "https://github.com/schroef/Save-Render-Presets/issues",
+    "wiki_url": "https://github.com/schroef/Save-Render-Presets",
     "category": "Render",
 }