+++ eevee_presets.py
@@ -123,20 +123,23 @@ class EEVEE_OT_AddEeveePreset(AddPresetBase, Operator):
             line = f"{key} = {value}\n"
             preset_lines.append(line)
 
-        user_path = Path(bpy.utils.resource_path('USER'))
-        preset_path = user_path / Path(f"scripts/presets/{PRESET_SUBDIR}")
+        # from preset.py
+        target_path = os.path.join("presets", PRESET_SUBDIR)
+        preset_path = bpy.utils.user_resource('SCRIPTS',target_path,create=True)
+        print(preset_path)
 
-        try:
-            if not preset_path.exists():
-                preset_path.mkdir()
-        except PermissionError as _:
-            self.report(
-                {'ERROR'}, f"PermissionError for '{preset_path}'")
-            return {'CANCELLED'}
-        except FileNotFoundError as _:
-            self.report(
-                {'ERROR'}, f"FileNotFoundError for '{preset_path}'")
-            return {'CANCELLED'}
+        # Causes error when using sub dirs
+        # try:
+        # if not preset_path.exists():
+        #     preset_path.mkdir()
+        # except PermissionError as _:
+        #     self.report(
+        #         {'ERROR'}, f"PermissionError for '{preset_path}'")
+        #     return {'CANCELLED'}
+        # except FileNotFoundError as _:
+        #     self.report(
+        #         {'ERROR'}, f"FileNotFoundError for '{preset_path}'")
+            # return {'CANCELLED'}
 
         filename = self.as_filename(self.preset_name)
         preset_file_path = preset_path / Path(f"{filename}.py")