+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 - Remove preset operator was not working
 - Adding new preset did not make it active in the preset menu
 
+### Changed
+- Info links Preferenes, changed warning to wiki_url and track_url
+
 ## [0.0.2] - 2021-10-14
 ### Changed
 - Code refactoring. Using Daniel Elther method of retrieving keys is more Blender version efficient. Does need manual attributes to be added, works on different version without effort, and thus more efficient