+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.4] - 2022-08-15
+### Fixed
+- Fixed issue when render-preset folder is created first time with cycles or eevee
+
 ## [0.0.3] - 2021-11-16
 ### Fixed
 - Remove preset operator was not working
@@ -24,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.4]:https://github.com/schroef/Save-Render-Presets/releases/tag/v0.0.4
 [0.0.3]:https://github.com/schroef/Save-Render-Presets/releases/tag/v0.0.3
 [0.0.2]:https://github.com/schroef/Save-Render-Presets/releases/tag/v0.0.2
 [0.0.1]:https://github.com/schroef/Save-Render-Presets/releases/tag/v0.0.1